string to double, not int

+++ ocr.cpp
@@ -165,7 +165,7 @@ void test()	//TODO: MAKE SOME NEURAL NETS, READ THE WEIGHTS FROM A FILE perceptr
 	for (int i = 0; i < GRIDHEIGHT*GRIDWIDTH; i++)//set it up with all of the correct output weights
 	{
 		getline(perceptronFile, line);
-		neuron->outputweight[i] = std::stoi(line);
+		neuron->outputweight[i] = std::stod(line);
 		line = "";
 	}
 
@@ -174,7 +174,7 @@ void test()	//TODO: MAKE SOME NEURAL NETS, READ THE WEIGHTS FROM A FILE perceptr
 		for (int j = 0; j < GRIDHEIGHT*GRIDWIDTH + 1; j++)
 		{
 			getline(perceptronFile, line);
-			neuron->hiddenweight[i][j] = std::stoi(line);
+			neuron->hiddenweight[i][j] = std::stod(line);
 			line = "";
 
 		}
